move padding code to separate library, add matrix_half_plus_two_model for padding tests, add end-to-end padding test for batching session
