+++ tensorflow_serving/batching/BUILD
@@ -18,6 +18,12 @@ filegroup(
     ),
 )
 
+filegroup(
+    name = "matrix_half_plus_two",
+    srcs = glob(
+        ["testdata/matrix_half_plus_two/**/*"]),
+)
+
 cc_library(
     name = "streaming_batch_scheduler",
     srcs = ["streaming_batch_scheduler.cc"],
@@ -57,6 +63,7 @@ cc_library(
         "//tensorflow_serving/servables/tensorflow:serving_session",
         "//tensorflow_serving/util:cleanup",
         "//tensorflow_serving/util:hash",
+        "//tensorflow_serving/batching:batching_util",
         "@org_tensorflow//tensorflow/contrib/batching:basic_batch_scheduler",
         "@org_tensorflow//tensorflow/contrib/batching:batch_scheduler",
         "@org_tensorflow//tensorflow/core:core_cpu",
@@ -72,7 +79,7 @@ cc_test(
         "batching_session_test.cc",
     ],
     data = [
-        "@org_tensorflow//tensorflow/cc/saved_model:saved_model_half_plus_two",
+        "@org_tensorflow//tensorflow/cc/saved_model:saved_model_half_plus_two", ":matrix_half_plus_two"
     ],
     deps = [
         ":batching_session",
@@ -112,5 +119,30 @@ cc_test(
         "@org_tensorflow//tensorflow/core:lib",
         "@org_tensorflow//tensorflow/core:protos_all_cc",
         "@org_tensorflow//tensorflow/core:test",
+    ],
+)
+
+cc_library(
+    name = "batching_util",
+    srcs = ["batching_util.cc"],
+    hdrs = ["batching_util.h"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "@org_tensorflow//tensorflow/core:framework",
+        "@org_tensorflow//tensorflow/core:lib",
+        "@org_tensorflow//tensorflow/contrib/batching:basic_batch_scheduler",
+    ],
+)
+
+cc_test(
+    name = "batching_util_test",
+    srcs = [
+        "batching_util_test.cc"
+    ],
+    deps = [
+        ":batching_util",
+        "//tensorflow_serving/core/test_util:test_main",
+        "@org_tensorflow//tensorflow/core:framework",
+        "@org_tensorflow//tensorflow/core:lib",
     ],
 )
